move test to its own file
node-http.test.ts has an unrelated environment-dependent failure (the
issue#4295 proxy test's server listens on ::1 while the client resolves
localhost to 127.0.0.1 in some containers); keep the new coverage in a
self-contained file so it runs clean everywhere.
